Move source root to repo root.
This is a more natural example for current Python repos.
