+++ BUILD
@@ -6,11 +6,11 @@
 # with the name of the corresponding dist.
 #
 # For example, `translate>=3.2.1` expands to:
+#   python_requirement_library(
+#     name='translate',
+#     requirements=[
+#       python_requirement('translate>=3.2.1')
+#     ]
+#   )
 
-# python_requirement_library(
-#   name='translate',
-#   requirements=[
-#     python_requirement('translate>=3.2.1')
-#   ]
-# )
 python_requirements()
