+++ helloworld/greet/greeting_test.py
@@ -1,7 +1,7 @@
 # Copyright 2020 Pants project contributors.
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from helloworld.greeting import Greeter
+from helloworld.greet.greeting import Greeter
 
 
 def test_greeter():
